backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.26.0
aiohttp==3.9.3
aiosqlite==0.20.0
sqlalchemy==2.0.27
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.2.0
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
python-dotenv==1.0.1
yfinance==0.2.37
requests==2.31.0
python-dateutil==2.8.2
apscheduler==3.10.4
